+++ CHANGELOG.md
@@ -1137,3 +1137,4 @@
 - [2026-05-14 11:46] feat(state): add bi-directional event bus handler for state
 - [2026-05-14 12:05] feat(forms): add bi-directional event bus handler for forms
 - [2026-05-14 12:13] fix(state): fix memory leak by cancelling state subscription
+- [2026-05-14 12:32] feat(security): implement shimmer loading skeleton in security